Refactor getClients method to improve code readability and maintainability
